Fix blocking opcode for keypress.
Opcode 0xFX0A is a blocking operation, that waits till a key has been pressed on the keypad.
The straightforward way to implement this, is an infinite loop. If you do this, then the rest of the emulator will also be blocked (including input handling).

A better method is by checking the keypad once, and only incrementing the program counter if a key has been pressed.
This way, the program counter will be stuck on the same opcode, till a key has been pressed.